FIX: time for pause/refresh intervall now moved to sync_step() => handled internally each refresh => better responsivity FIX: end of station with layout>4 in NS direction
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@728 8aca7d54-2c30-db11-9de9-000461428c89

+++ simcity.cc
@@ -372,7 +372,7 @@ bool stadt_t::cityrules_init(cstring_t objfilename)
 
 		// find out rule size
 		int size = 0;
-		int maxlen = strlen(rule);
+		size_t maxlen = strlen(rule);
 		while (size < maxlen && rule[size] != ' ') size++;
 
 		if (size > 7 || maxlen < size * (size + 1) - 1 || (size & 1) == 0 || size <= 2) {
@@ -402,7 +402,7 @@ bool stadt_t::cityrules_init(cstring_t objfilename)
 
 		// find out rule size
 		int size = 0;
-		int maxlen = strlen(rule);
+		size_t maxlen = strlen(rule);
 		while (size < maxlen && rule[size] != ' ') size++;
 
 		if (size > 7 || maxlen < size * (size + 1) - 1 || (size & 1) == 0 || size <= 2) {
@@ -623,7 +623,7 @@ stadt_t::haltestellenname(koord k, const char *typ, int number)
 		}
 	}
 
-	const long len = strlen(buf) + 1;
+	const size_t len = strlen(buf) + 1;
 	char* name = new char[len];
 	tstrncpy(name, buf, len);
