+++ utils/simstring.c
@@ -206,7 +206,7 @@ char *make_single_line_string(const char *in,int number_of_lines)
  * @return dest
  * @author Hj. Malthaner
  */
-char *tstrncpy(char *dest, const char *src, int n)
+char *tstrncpy(char *dest, const char *src, size_t n)
 {
   strncpy(dest, src, n);
   dest[n-1] = '\0';
